+++ crates/ruma-common/src/push.rs
@@ -39,6 +39,7 @@ pub use self::{
     action::{Action, Tweak},
     condition::{
         ComparisonOperator, FlattenedJson, PushCondition, PushConditionRoomCtx, RoomMemberCountIs,
+        _CustomPushCondition,
     },
     iter::{AnyPushRule, AnyPushRuleRef, RulesetIntoIter, RulesetIter},
     predefined::{